Use rayon for work-stealing
This improves the concurrency of our Mandelbrot program. The pale gray
parts of the image, where the loop quickly exits, are much quicker to
render than the black parts (which do the full 255 iterations). So
splitting the bands into equal sizes does not mean the amount of work
done on each band is equal.

We introduce Rayon here to fire off a parallel task for each row of the
pixels in the output. This creates several hundred tasks that Rayon can
distribute across its internal thread pool. Thanks to work-stealing, it
doesn't matter that the tasks require a different amount of computation.
Rayon will balance the work as it goes.

On my 2017 MBP with 3.3 GHz Dual-Core Intel Core i5, we can see before
and after behaviour.

Starting point:

```sh
>  time ./target/release/mandelbrot mandel.png 4000x3000 -1.20,0.35 -1,0.20
4.33s user 0.04s system 90% cpu 4.804 total
```

With Rayon:
```sh
>  time ./target/release/mandelbrot mandel.png 4000x3000 -1.20,0.35 -1,0.20
4.45s user 0.03s system 327% cpu 1.367 total
```